Handle variadic functions just like normal functions
ctypes simply supports assigning argtypes to variadic functions, and
typechecks all non-variadic arguments.

The ctypes docs note that specifying argtypes on variadic functions is
mandatory on macOS arm64, because the calling convention differs from
that of regular functions. [1]

It is not mentioned if argtypes support was introduced in a specific
python version.
However, the docs say "Because specifying the attribute does not inhibit
portability it is advised to always specify argtypes for all variadic
functions."

[1]
https://docs.python.org/3/library/ctypes.html#calling-variadic-functions