Do allow c_void_p as restype
AFAICS the autocast should be fine. A c_void_p is just that: an int, or
None for a null pointer.
I don't think it's problematic, just different (direct access rather
than through .value).
Note that you can seamlessly ctypes.cast() an int to a pointer type.

Generally, correctness of ctypes is much higher than correctness of
legacy ctypesgen code.
In the past, most wrappers have turned out to be unnecessary or
incorrect. The problem is that some of them were extremely difficult to
deal with to their bloated or obscure nature, which presumably is
why noone dared touch them previously.