Remove clerical files we don't use/maintain
* Debian packaging belongs in a separate repo. To my knowledge, noone is
currently using or testing this code (not even upstream).
* We're not interested in flake8, black, tox or travis
* The contributing rules are trivial/obsolete
* We don't maintain the changelog. The info of what has happened in the
past is nice but still viewable upstream or in the git history.
* todo.txt hasn't been touched for over 16 years and its content doesn't
seem overly relevant
* docs/publishing is much text but of little value to us. It mostly
discusses versioning, which we plan to revise anyway. Common
versioning/publishing guidelines for python projects are documented
elsewhere or are trivial to anyone who has been working in the python
world for some time.