preamble/ptrdiff: order int types by likeliness
We are most likely on a platform that uses 64-bit pointers.

Expectably, 16-bit pointers are not a realistic target for ctypesgen,
but retain it anyways to be C standard correct.

Note, the point of this template is to achieve universal bindings. If
bindings were target platform specific, the ptrdiff_t constant pulled in
from stddef.h would be fine.