Somewhat allow linking relative modules from higher dirs
i.e. relative imports with more than 1 leading dot

However, the limitation here is, if we are operating relative to a
module that is not top-level, it must be self-contained, i.e. the import
tree is not permitted to ascend beyond the anchor.

In other words, guessing the anchor from outpath has limitations.
A possible solution might be to take the higher anchor from the caller.
In that case, we'd need to transform the compile-time import path, but
it should be possible to deduce this programatically from the number of
dots and the diff between default anchor and higher anchor.