Check FAM after processing
The processed ctype object gives us a slightly higher-level interface,
which is cleaner to work with.

Also mention that this is the step just before the printer (i.e. in
principle, we could also do this in print_struct(), as is currently the
case with packed/unnamed, but this is better, as it should apply for all
printers, not just the python one).