pp: add whitespace normalization to commented-out section
Dispose of whitespace between hash and specifier that causes the lexer
to fail. For this purpose, expand the regex to capture the remainder of
the line.

Also change the regex to recognize any kind of whitespace after the
specifier, not just a space, to avoid errors like:
```
WARNING: Skip unhandled directive 'define\t__LITTLE_ENDIAN\t1234\n'
```

Good pre-processors like gcc or clang should normalize this in the first
place. However, pcpp currently does not.

---

Also, remove excessive comments after all, it's clear enough in
retrospect.