--no-{stddef,gnu-python}-types: internal bool normalization
Internally, these flags are checked before adding types, so it makes
more sense to use the positive condition under the hood, which is also
the default.
Otherwise, you would nearly always have to prepend a `not`, which is
distasteful.

The front-end stays the same.
These options could also be converted to `BooleanOptionalAction`s if
desired, but I did not see an imminent need for that.