ctypesdescs: normalize code order in helper function
Writing accessing code before the variables are defined doesn't make for
good readability, though it worked because the real access is deferred
until the actual method calls.

Also, acknowledge that this kind of closure seems a bit distasteful.
It might make more sense to create a proper initializer and store these
on object level, rather than declaring a class inside a function.