Replace visit_type_and_collect_info() with CollectingTypeVisitor
There is only one caller of visit_type_and_collect_info(), which
iterates over objects and merges the results into lists, so it is
straightforward to unroll the procedure and skip a series of unnecessary
list constructions and `.extend()`s.

It may well be possible to use the same visitor object for all
iterations, but I haven't thoroughly checked that yet.