Maintain indentation levels in most sources
This makes it possible to paste ctypesgen code into a python console.
Also, it makes editing a bit more convenient.

If - for whatever reason - you do not want the additional whitespace,
use a tool to remove it when installing/packaging ctypesgen.
But the actual source repo should keep the indents.

Skipped files: cparser, cgrammar, cdeclarations because we barely touch
these. If we do at some point, these can be converted too.