Avoid free_library() after all
Using the internal _ctypes isn't ideal and adds fragility.
What's more, this is only ever needed in a very niche case (API callers
on Windows who want to delete the DLL in same session).

On Linux/macOS, we can drop this entirely. On Windows, use a combination
of free_library() [runtime] and --no-load-library [bindgen time] in test
suite.